Add cli.
